+++ CHANGELOG.rst
@@ -4,8 +4,9 @@ Version 3.3.0
 -------------
 * Type hints and stubs
 * PEP0518
-* Deprecation of *args for logwrap
-* Fix empty *args and **kwargs
+* Deprecation of `*args` for logwrap
+* Fix empty `*args` and `**kwargs`
+* allow override for arguments repr processing (pre- and post-processing)
 
 Version 3.2.0
 -------------